requirements.txt
anyio==2.2.0
asgiref==3.3.4
requests==2.25.1
click==7.1.2 # RQ worker пока не работает с click версии выше
Django==3.2.2
rollbar==0.15.2
environs[django]==9.3.2
psycopg2-binary==2.8.5
Pillow==8.2.0
phonenumbers==8.12.22
django-phonenumber-field==5.1.0
dbr==8.2
django-rq==2.4.1
djangorestframework==3.12.4
rq-dashboard==0.6.1
more-itertools==8.7.0 # пока что временно 
